+++ src/Search/Repository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace srag\Plugins\SrSearchPDBlock\Block;
+namespace srag\Plugins\SrSearchPDBlock\Search;
 
 use ilSrSearchPDBlockPlugin;
 use srag\DIC\SrSearchPDBlock\DICTrait;
@@ -9,7 +9,7 @@ use srag\Plugins\SrSearchPDBlock\Utils\SrSearchPDBlockTrait;
 /**
  * Class Repository
  *
- * @package srag\Plugins\SrSearchPDBlock\Block
+ * @package srag\Plugins\SrSearchPDBlock\Search
  *
  * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  */
@@ -63,6 +63,35 @@ final class Repository
     public function factory() : Factory
     {
         return Factory::getInstance();
+    }
+
+
+    /**
+     * @param string $key_config_global_search
+     * @param string $key_config_current_page_search
+     * @param bool   $everywhere
+     *
+     * @return string
+     */
+    public function getSearches(string $key_config_global_search, string $key_config_current_page_search, bool $everywhere = false) : string
+    {
+        $searches = [];
+
+        if (self::srSearchPDBlock()->config()->getValue($key_config_global_search)) {
+            $searches[] = $this->factory()->globalSearch();
+        }
+
+        if (self::srSearchPDBlock()->config()->getValue($key_config_current_page_search)) {
+            $searches[] = $this->factory()->currentPageSearch();
+        }
+
+        if ($everywhere) {
+            $searches = $this->factory()->everywhereSearch($searches);
+        } else {
+            $searches = array_map([$this->factory(), "searchBlock"], $searches);
+        }
+
+        return self::output()->getHTML($searches);
     }